feat: validate urls in parallel
This speeds up the code by quite a bit. It's not the cleanest since
we're still awaiting the urls in order, but that doesn't matter that
much since the requests themselves are made in parallel, and actually
collecting the errors is a small portion of the time spent.
I tried to use tokio streams but since we can't use async blocks inside
closures to use combinators like `map` or `filter` I couldn't figure out
a cleaner way of doing this